+++ app/screens/detailTransaction/hooks/useDetailTransaction.ts
@@ -1,9 +1,11 @@
+import { useContext } from "react";
 import { Alert } from "react-native";
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
 
 import RootStackParamList from "../../../models/interfaces/RootScreensParams";
 
 import { RealmContext } from "../../../configs/RealmContext";
+import { AppConfigContext } from "../../../context/appConfig.context";
 import * as transactionService from '../../../services/transaction.service';
 
 const { useRealm } = RealmContext;
@@ -12,6 +14,7 @@ const useDetailTransaction = ({ navigation, route }: NativeStackScreenProps<Root
   const { transaction } = route.params;
 
   const realm = useRealm();
+  const { appConfig } = useContext(AppConfigContext);
 
   const deleteTransaction = () => {
     try {
@@ -31,6 +34,7 @@ const useDetailTransaction = ({ navigation, route }: NativeStackScreenProps<Root
   }
   
   return {
+    appConfig,
     transaction,
     onDeleteButtonPress
   }